requirements.txt
# Core Dependencies
boto3>=1.34.0              # S3/R2 client
qdrant-client>=1.7.0       # Qdrant operations
langchain>=0.1.0           # Chunking & text splitting
tiktoken>=0.5.0            # Token counting
requests>=2.31.0           # HTTP requests
python-dotenv>=1.0.0       # Environment variables
pydantic>=2.5.0            # Data validation
tenacity>=8.2.0            # Retry logic
xxhash>=3.4.0              # Lightweight hashing

# FastAPI for n8n integration
fastapi>=0.104.0           # API framework
uvicorn>=0.24.0            # ASGI server

# Development & Testing
pytest>=7.4.0              # Testing framework
pytest-cov>=4.1.0          # Coverage reporting
black>=23.0.0              # Code formatting
flake8>=6.0.0              # Linting
